Add files via upload
This dataset provides the annual inflation percentage in the United States from the year 1960 to 2022. Each entry includes the year (in the format YYYY-MM-DD) and the corresponding inflation percentage. Inflation percentage indicates the rate at which the general level of prices for goods and services is rising, leading to a decrease in purchasing power over time. The data can be utilized for analyzing historical trends, forecasting future economic conditions, and making informed decisions in various financial and economic contexts.